dj/pyproject.toml
[tool.poetry]
name = "dj"
version = "0.1.0"
description = ""
authors = ["sximada <50688746+sximada@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
django-cors-headers = "^3.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
